+++ src/Exceptionist.php
@@ -23,6 +23,7 @@ use Tools\Exception\FileNotExistsException;
 use Tools\Exception\KeyNotExistsException;
 use Tools\Exception\NotReadableException;
 use Tools\Exception\PropertyNotExistsException;
+use Tools\Filesystem;
 
 /**
  * Exceptionist.
@@ -114,7 +115,7 @@ class Exceptionist
      */
     public static function fileExists(string $filename, $message = '', $exception = FileNotExistsException::class): string
     {
-        $message = $message ?: sprintf('File or directory `%s` does not exist', rtr($filename));
+        $message = $message ?: sprintf('File or directory `%s` does not exist', (new Filesystem())->rtr($filename));
         self::isTrue(file_exists($filename), $message, $exception);
 
         return $filename;
@@ -134,7 +135,7 @@ class Exceptionist
     {
         self::fileExists($filename, $message, $exception);
 
-        $message = $message ?: sprintf('File or directory `%s` is not readable', rtr($filename));
+        $message = $message ?: sprintf('File or directory `%s` is not readable', (new Filesystem())->rtr($filename));
         self::isTrue(is_readable($filename), $message, $exception);
 
         return $filename;
@@ -154,7 +155,7 @@ class Exceptionist
     {
         self::fileExists($filename, $message, $exception);
 
-        $message = $message ?: sprintf('File or directory `%s` is not writable', rtr($filename));
+        $message = $message ?: sprintf('File or directory `%s` is not writable', (new Filesystem())->rtr($filename));
         self::isTrue(is_writable($filename), $message, $exception);
 
         return $filename;